+++ src/main/java/roomescape/config/WebMvcConfig.java
@@ -1,28 +1,28 @@
-package roomescape.web.config;
+package roomescape.config;
 
 import java.util.List;
-import org.springframework.context.annotation.Bean;
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.application.security.JwtProvider;
+import roomescape.web.support.AdminHandlerInterceptor;
+import roomescape.web.support.LoginMemberArgumentResolver;
 
 @Configuration
-public class WebMvcConfiguration implements WebMvcConfigurer {
-    @Bean
-    public AdminHandlerInterceptor adminHandlerInterceptor() {
-        return new AdminHandlerInterceptor(new JwtProvider());
-    }
+@RequiredArgsConstructor
+public class WebMvcConfig implements WebMvcConfigurer {
+    private final AdminHandlerInterceptor adminHandlerInterceptor;
+    private final LoginMemberArgumentResolver loginMemberArgumentResolver;
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new LoginMemberArgumentResolver());
+        resolvers.add(loginMemberArgumentResolver);
     }
 
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(adminHandlerInterceptor())
+        registry.addInterceptor(adminHandlerInterceptor)
                 .addPathPatterns("/admin/**");
     }
 }
