+++ src/main/java/roomescape/web/support/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.web.config;
+package roomescape.web.support;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
